# Optimized tip_tip_distances