Significantly improved the performance of `TreeNode.tip_tip_distances`.

This method generates a distance matrix of the sum of branch lengths connecting each pair of tips in a tree. It is a prerequisite for calculating some tree-to-tree distance metrics, such as path-length distance. The previous code was fairly efficient. This PR further improved the performance by maximizing the use of NumPy vectorization, making tip index ranges continuous (therefore becoming slices), and removing the need for an intermediate square matrix of the total number of tips in the tree, which is a major bottleneck, especially when only a subset of the tips are to be calculated.

Test on a tree with 10,575 tips:

- All tips
  - Before: 12.2s
  - After: 318 ms
- 5,000 randomly selected tips
  - Before: 3.59 s
  - After: 254 ms

***

Please complete the following checklist:

* [x] I have read the [contribution guidelines](https://scikit.bio/contribute.html).

* [x] I have documented all public-facing changes in the [changelog](https://github.com/scikit-bio/scikit-bio/blob/main/CHANGELOG.md).

* [ ] **This pull request includes code, documentation, or other content derived from external source(s).** If this is the case, ensure the external source's license is compatible with scikit-bio's [license](https://github.com/scikit-bio/scikit-bio/blob/main/LICENSE.txt). Include the license in the `licenses` directory and add a comment in the code giving proper attribution. Ensure any other requirements set forth by the license and/or author are satisfied.

  - **It is your responsibility to disclose** code, documentation, or other content derived from external source(s). If you have questions about whether something can be included in the project or how to give proper attribution, include those questions in your pull request and a reviewer will assist you.

* [x] This pull request does not include code, documentation, or other content derived from external source(s).

Note: [This document](https://scikit.bio/devdoc/review.html) may also be helpful to see some of the things code reviewers will be verifying when reviewing your pull request.